widgets/views.py
from rest_framework import viewsets
from .models import Widget
from .serializers import WidgetSerializer

class WidgetViewSet(viewsets.ModelViewSet):
    queryset = Widget.objects.all()
    serializer_class = WidgetSerializer